+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All important changes to `fiisoft-queue-provider` will be documented in this file
 
+## 1.2.0
+
+- class InstantCommandQueue operates now much much faster thanks to use SplQueue
+- version od dependency fiisoft/tasks-queue lifted to 2.0.0
+
 ## 1.1.0
 
 Added InstantCommandQueue - a new implementation of CommandQueue from fiisoft/tasks-queue